Rename the navigation folder to navigators. This will allow it to play nicer with the ignite cli to generate navigators.
