+++ boilerplate/app/services/reactotron/reactotron.ts
@@ -5,7 +5,7 @@ import { onSnapshot } from "mobx-state-tree"
 import { ReactotronConfig, DEFAULT_REACTOTRON_CONFIG } from "./reactotron-config"
 import { mst } from "reactotron-mst"
 import { clear } from "../../utils/storage"
-import { RootNavigation } from "../../navigation"
+import { RootNavigation } from "../../navigators"
 
 // Teach TypeScript about the bad things we want to do.
 declare global {
