+++ docs/Folder-Structure.md
@@ -10,7 +10,7 @@ ignite-project
 │   ├── components
 │   ├── i18n
 │   ├── models
-│   ├── navigation
+│   ├── navigators
 │   ├── screens
 │   ├── services
 │   ├── theme
@@ -49,7 +49,7 @@ app
 │── components
 │── i18n
 ├── models
-├── navigation
+├── navigators
 ├── screens
 ├── services
 ├── theme
@@ -69,7 +69,7 @@ This is where your translations will live if you are using `react-native-i18n`.
 
 This is where your app's models will live. Each model has a directory which will contain the `mobx-state-tree` model file, test file, and any other supporting files like actions, types, etc. There's also an extensions directory with useful shared extensions that you can include in your models like `.extend(withRootStore)` or `.extend(withEnvironment)` to access the root store or environment respectively.
 
-**navigation**
+**navigators**
 
 This is where your `react-navigation` navigators will live.
